requirements/base.txt
Django==2.1.3
django-appconf==1.0.2
django-betterforms==1.2
django-debug-toolbar==1.10.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
markdown2==2.3.6
Pillow==5.3.0
pip-autoremove==0.9.1
psycopg2-binary==2.7.6.1
python-magic==0.4.15
pytz==2018.7
sqlparse==0.2.4
URLObject==2.4.3
